Refactor decision cleanup into orchestrator
